Added constructor + call example
When Product is invoked, javascript creates a new object, with a prototype of Product.prototype  and then calls the Product function.

lets the Food constructor do all the same stuff that the Product constructor does, plus a bit of extra stuff. It's a way to reuse the code you wrote for Product, while adding specialized behavior that only matters to Food items.